server/requirements.txt
fastapi
uvicorn
psycopg2-binary
pika
pydantic-settings